lemon-utils/crypto-account-creator/requirements.txt
pandas==1.4.1
requests==2.27.1
tqdm==4.63.0
python-dotenv==0.19.2